fix : refactored HourRecap
