+++ app/src/test/java/com/android/sample/ui/calendar/CalendarViewModelTest.kt
@@ -141,104 +141,4 @@ class CalendarViewModelTest {
     viewModel.clearErrorMsg()
     assertNull(viewModel.uiState.value.errorMsg)
   }
-
-  @Test
-  fun calculateWorkedHoursShouldCalculateCorrectlyForPastEvents() = runTest {
-    // Past event (relative to now)
-    // We need to ensure the event is considered "past" by the ViewModel.
-    // Since ViewModel uses Instant.now(), and we can't easily mock it without dependency injection
-    // change,
-    // we will create an event very far in the past.
-    val pastStart = Instant.parse("2000-01-01T10:00:00Z")
-    val pastEnd = Instant.parse("2000-01-01T12:00:00Z") // 2 hours
-
-    val pastEvent =
-        createEvent(
-            organizationId = orgId,
-            title = "Past Meeting",
-            startDate = pastStart,
-            endDate = pastEnd,
-            participants = setOf("user1", "user2"),
-            presence = mapOf("user1" to true, "user2" to false))[0]
-
-    repositoryEvent.insertEvent(orgId, pastEvent)
-
-    viewModel.calculateWorkedHours(pastStart.minusSeconds(1), pastEnd.plusSeconds(1))
-    testDispatcher.scheduler.advanceUntilIdle()
-
-    val result = viewModel.uiState.value.workedHours
-
-    // user1: 2 hours, user2: 0 hours (not present)
-    val user1Hours = result.find { it.first == "user1" }?.second ?: 0.0
-    val user2Hours = result.find { it.first == "user2" }?.second ?: 0.0
-
-    assertEquals(2.0, user1Hours, 0.01)
-    assertEquals(0.0, user2Hours, 0.01)
-  }
-
-  @Test
-  fun calculateWorkedHoursShouldCalculateCorrectlyForFutureEvents() = runTest {
-    // Future event (relative to now)
-    val futureStart = Instant.now().plusSeconds(3600) // 1 hour from now
-    val futureEnd = futureStart.plusSeconds(7200) // 2 hours duration
-
-    val futureEvent =
-        createEvent(
-            organizationId = orgId,
-            title = "Future Meeting",
-            startDate = futureStart,
-            endDate = futureEnd,
-            participants = setOf("user1", "user2"),
-            presence = emptyMap() // Presence shouldn't matter for future events
-            )[0]
-
-    repositoryEvent.insertEvent(orgId, futureEvent)
-
-    viewModel.calculateWorkedHours(futureStart.minusSeconds(1), futureEnd.plusSeconds(1))
-    testDispatcher.scheduler.advanceUntilIdle()
-
-    val result = viewModel.uiState.value.workedHours
-
-    // user1: 2 hours, user2: 2 hours (assumed present)
-    val user1Hours = result.find { it.first == "user1" }?.second ?: 0.0
-    val user2Hours = result.find { it.first == "user2" }?.second ?: 0.0
-
-    assertEquals(2.0, user1Hours, 0.01)
-    assertEquals(2.0, user2Hours, 0.01)
-  }
-
-  @Test
-  fun calculateWorkedHoursShouldAggregateHoursFromMultipleEvents() = runTest {
-    val start = Instant.parse("2000-01-01T00:00:00Z")
-    val end = Instant.parse("2000-01-02T00:00:00Z")
-
-    val event1 =
-        createEvent(
-            organizationId = orgId,
-            title = "Event 1",
-            startDate = Instant.parse("2000-01-01T10:00:00Z"),
-            endDate = Instant.parse("2000-01-01T11:00:00Z"), // 1 hour
-            participants = setOf("user1"),
-            presence = mapOf("user1" to true))[0]
-
-    val event2 =
-        createEvent(
-            organizationId = orgId,
-            title = "Event 2",
-            startDate = Instant.parse("2000-01-01T14:00:00Z"),
-            endDate = Instant.parse("2000-01-01T16:00:00Z"), // 2 hours
-            participants = setOf("user1"),
-            presence = mapOf("user1" to true))[0]
-
-    repositoryEvent.insertEvent(orgId, event1)
-    repositoryEvent.insertEvent(orgId, event2)
-
-    viewModel.calculateWorkedHours(start, end)
-    testDispatcher.scheduler.advanceUntilIdle()
-
-    val result = viewModel.uiState.value.workedHours
-
-    val user1Hours = result.find { it.first == "user1" }?.second ?: 0.0
-    assertEquals(3.0, user1Hours, 0.01)
-  }
 }
